- Added <Plug>(vinarise_change_current_address) mapping.
- Check modified flag.

+++ autoload/vinarise/mappings.vim
@@ -1,7 +1,7 @@
 "=============================================================================
 " FILE: mappings.vim
 " AUTHOR: Shougo Matsushita <Shougo.Matsu@gmail.com>
-" Last Modified: 14 Feb 2012.
+" Last Modified: 16 Feb 2012.
 " License: MIT license  {{{
 "     Permission is hereby granted, free of charge, to any person obtaining
 "     a copy of this software and associated documentation files (the
@@ -41,6 +41,8 @@ function! vinarise#mappings#define_default_mappings()"{{{
         \ :<C-u>call <SID>next_half_screen()<CR>
   nnoremap <buffer><silent> <Plug>(vinarise_print_current_position)
         \ :<C-u>call <SID>print_current_position()<CR>
+  nnoremap <buffer><silent> <Plug>(vinarise_change_current_address)
+        \ :<C-u>call <SID>change_current_address()<CR>
   "}}}
 
   if exists('g:vimshell_no_default_keymappings') && g:vimshell_no_default_keymappings
@@ -51,10 +53,11 @@ function! vinarise#mappings#define_default_mappings()"{{{
   nmap <buffer> V <Plug>(vinarise_edit_with_vim)
   nmap <buffer> q <Plug>(vinarise_hide)
   nmap <buffer> Q <Plug>(vinarise_exit)
-  nmap <buffer>j         <Plug>(vinarise_next_line)
-  nmap <buffer><C-f>     <Plug>(vinarise_next_screen)
-  nmap <buffer><C-d>     <Plug>(vinarise_next_half_screen)
-  nmap <buffer><C-g>     <Plug>(vinarise_print_current_position)
+  nmap <buffer> j         <Plug>(vinarise_next_line)
+  nmap <buffer> <C-f>     <Plug>(vinarise_next_screen)
+  nmap <buffer> <C-d>     <Plug>(vinarise_next_half_screen)
+  nmap <buffer> <C-g>     <Plug>(vinarise_print_current_position)
+  nmap <buffer> r    <Plug>(vinarise_change_current_address)
 endfunction"}}}
 
 " VimShell key-mappings functions.
@@ -69,14 +72,32 @@ function! s:edit_with_vim()"{{{
   endtry
 endfunction"}}}
 function! s:hide()"{{{
+  if b:vinarise.modified
+    let yes = input(
+          \ 'Current vinarise buffer is modified! Hide anyway?: ', 'yes')
+    redraw
+    if yes !~ '^y\%[es]$'
+      return
+    endif
+  endif
+
   " Switch buffer.
   if winnr('$') != 1
-    close
+    close!
   else
     call vinarise#util#alternate_buffer()
   endif
 endfunction"}}}
 function! s:exit()"{{{
+  if b:vinarise.modified
+    let yes = input(
+          \ 'Current vinarise buffer is modified! Exit anyway?: ')
+    redraw
+    if yes !~ '^y\%[es]$'
+      return
+    endif
+  endif
+
   call vinarise#util#delete_buffer()
 endfunction"}}}
 function! s:print_current_position()"{{{
@@ -85,6 +106,38 @@ function! s:print_current_position()"{{{
         \ vinarise#get_cur_text(getline('.'), col('.')))
   echo printf('[%s] %8d / %8d byte (%3d%%)',
         \ type, address, b:vinarise.filesize, (address*100)/b:vinarise.filesize)
+endfunction"}}}
+function! s:change_current_address()"{{{
+  " Get current address.
+  let [type, address] = vinarise#parse_address(getline('.'),
+        \ vinarise#get_cur_text(getline('.'), col('.')))
+  if type == 'address'
+    " Invalid.
+    return
+  endif
+
+  execute 'python' 'vim.command("let old_value = " + str('.
+        \ b:vinarise.python .'.get_byte(vim.eval("address"))))'
+
+  let value = input('Please input new value: '.
+        \ printf('%x', old_value) . ' -> ')
+  redraw
+  if value == '' || value !~ '^\x\x\?$'
+    echo 'Invalid value.'
+    return
+  endif
+  let value = str2nr(value, 16)
+
+  execute 'python' b:vinarise.python .
+        \ '.set_byte(vim.eval("address"), vim.eval("value"))'
+
+  setlocal modifiable
+
+  " Change current line.
+  call setline('.', vinarise#make_line(address / 16))
+  let b:vinarise.modified = 1
+
+  setlocal nomodifiable
 endfunction"}}}
 
 function! s:next_line()"{{{
